requirements.txt
# Uptime Returns System Requirements
# Python 3.11+

# Core Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
pyodbc==5.0.1
pymssql==2.2.11

# Email & OAuth
msal==1.33.0
requests==2.32.5

# Security (optional - for authentication)
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Utilities
python-dotenv==1.0.0
aiofiles==23.2.1

# Development (optional)
pytest==7.4.3
black==23.11.0
